Allow import strategy to update images and variant data
I based my import strategy on the default medusa import batch job. While updating products, variants' data was not updated because the update method returned before saving the variant. This also prevented product.updated events from triggering, which probably causes issues with search indexes that need this event to reindex data.

I tracked things down to this service. The changes made it possible for the images and variant data to be updated on import.

What I am not sure yet is how this acts when a variant is created while importing, but in this case, the import can simply be run twice for images to be attached.

Another issue that I have is that imported images can't be removed from a variant. They are not listed in the modal. But since I rely primarily on imports to populate data and the source of truth is elsewhere, this is good enough for me.